pulled TripleTriad into it's own file
